requirements.txt
anyjson==0.3.3
asgiref==3.2.7
attrs==19.3.0
coverage==5.1
Django==3.0.6
django-braces==1.14.0
django-model-utils==4.0.0
django-star-ratings==0.9.1
django-utils2==2.8.0
gunicorn==20.0.4
mock==4.0.2
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytest-django==3.9.0
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
wcwidth==0.1.9
